% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora.R
\name{.ora}
\alias{.ora}
\title{Perform ORA}
\usage{
.ora(
  dea_res,
  enrich_fun,
  result_class,
  dea_p_cutoff = 0.05,
  dea_log2fc_cutoff = c(-1, 1),
  ...
)
}
\arguments{
\item{dea_res}{DEA result from glystats.}

\item{enrich_fun}{An enrichment function.}

\item{result_class}{A string of the concrete result class.}

\item{dea_p_cutoff}{P-value cutoff to define statistical significance.}

\item{dea_log2fc_cutoff}{Log2FC cutoffs to define statistical significance.}

\item{...}{Parameters passed to downstream enrichment package.}
}
\description{
Perform ORA
}
