% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc-ora.R
\name{enrich_gc_ora_do}
\alias{enrich_gc_ora_do}
\title{Glycan-Centric Disease Ontology (DO) Over Representation Analysis}
\usage{
enrich_gc_ora_do(
  dea_res,
  dea_p_cutoff = 0.05,
  dea_log2fc_cutoff = c(-1, 1),
  ont = "HDO",
  organism = "hsa",
  universe = NULL,
  p_adj_method = "BH",
  p_cutoff = 0.05,
  q_cutoff = 0.2,
  min_gs_size = 10,
  max_gs_size = 500
)
}
\arguments{
\item{dea_res}{Differential analysis result. Can be one of:
\itemize{
\item Result from \code{\link[glystats:gly_limma]{glystats::gly_limma()}} (two groups), \code{\link[glystats:gly_ttest]{glystats::gly_ttest()}}, or \code{\link[glystats:gly_wilcox]{glystats::gly_wilcox()}},
called on an \code{\link[glyexp:experiment]{glyexp::experiment()}} of "traitproteomics" type.
\item A tibble with the following columns:
\itemize{
\item \code{protein}: Uniprot ID of proteins
\item \code{trait}: A glycosylation trait (e.g. "TFc" for proportion of core-fucosylated glycans)
\item \code{site}: The glycosylation site.
\item \code{p_val}: p-values, preferably adjusted p-values
\item \code{log2FC}: log2 of fold change
}
}}

\item{dea_p_cutoff}{P-value cutoff for statistical significance. Defaults to 0.05.
For \code{glystats} result input, adjusted p-values are used.}

\item{dea_log2fc_cutoff}{Log2 fold change cutoff statistical significance.
A length-2 numeric vector, being negative and positive boundaries, respectively.
For example, \code{c(-1, 1)} means "log2FC < -1 or log2FC > 1", and \code{c(-Inf, 1)} means "log2FC > 1".
Defaults to \code{c(-1, 1)}.}

\item{ont}{One of "HDO" (Human Disease Ontology), "MPO" (Mammalian Phenotype Ontology),
or "VDO" (Vector Disease Ontology). Passed to \code{ont} of \code{\link[DOSE:enrichDO]{DOSE::enrichDO()}}.
Defaults to "HDO".}

\item{organism}{"hsa" (Homo sapiens) or "mmu" (Mus musculus).
Passed to \code{organism} of \code{\link[DOSE:enrichDO]{DOSE::enrichDO()}}. Defaults to "hsa".}

\item{universe}{Background genes Uniprot IDs, directly passed to \code{universe} of downstream enrichment function.
If \code{NULL} (default), all genes in the data will be used.
Another common pattern is to use all detected proteins as backgroud genes.
You can use \code{\link[=detected_universe]{detected_universe()}} to help you.}

\item{p_adj_method}{P-value adjustment method.
One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".
Passed to \code{pAdjustMethod} of downstream enrichment function.
Defaults to "BH".}

\item{p_cutoff}{P-value cutoff to filter significant terms.
Passed to \code{pvalueCutoff} of downstream enrichment function. Defaults to 0.05.}

\item{q_cutoff}{Q-value (FDR) cutoff to filter significant terms.
Passed to \code{qvalueCutoff} of downstream enrichment function. Defaults to 0.2.}

\item{min_gs_size}{Minimal size of each gene set for analyzing.
Gene sets with fewer genes than this threshold will be excluded.
Passed to \code{minGSSize} of downstream enrichment function. Defaults to 10.}

\item{max_gs_size}{Maximum size of each gene set for analyzing.
Gene sets with more genes than this threshold will be excluded.
Passed to \code{maxGSSize} of downstream enrichment function. Defaults to 500.}
}
\value{
A clusterProfiler \code{compareClusterResult} object with additional \code{glyfun} classes.
It can be readily converted to a tibble with \code{\link[tibble:as_tibble]{tibble::as_tibble()}},
or visualized with \code{clusterProfiler} functions like \code{\link[clusterProfiler:reexports]{clusterProfiler::dotplot()}}.
}
\description{
Performs glycan-centric Disease Ontology (DO) Over-Representation Analysis (ORA).
Instead of traditional protein-centric enrichment, this function links specific
glycan traits to disease associations. It helps answer questions like "Which
diseases are enriched in proteins with a specific dysregulated glycan motif?",
by grouping differential analysis results by glycan traits and computing
disease enrichment for each trait.
}
\section{What is glycan-centric enrichment?}{
In traditional glycoproteomics data analysis,
we usually perform differential expression analysis (DEA) on glycoforms,
extract proteins that have dysregulated glycosylation,
then perform functional enrichment (e.g. GO) on these proteins.
This is what \code{enrich_xxx()} functions do (e.g. \code{\link[=enrich_ora_go]{enrich_ora_go()}}).

\code{enrich_gc_xxx()} functions differ in that they link specific glycan traits with functional annotations.
Instead of answering the question
"Which functions are enriched in dysregulated glycoproteins?",
\code{enrich_gc_xxx()} answers questions like
“Which functions are enriched in proteins with dysregulated core-fucosylation?”
Higher specificity, deeper insights. By focusing on distinct glycan motifs,
it helps you pinpoint the functional relevance of specific glycosylation changes.
}

\section{Common usage pattern}{
A common pattern of using this function is:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# 1. Use `glydet` to calculate derived traits or motif quantification.
trait_exp <- derive_traits(exp)  # or `quantify_motifs()`

# 2. Perform differential analysis with `glystats`.
dea_res <- gly_ttest(trait_exp)

# 3. Use this function.
go_res <- enrich_gc_ora_go(dea_res)  # or other `enrich_gc_xxx()` functions
}\if{html}{\out{</div>}}
}

\seealso{
\code{\link[clusterProfiler:compareCluster]{clusterProfiler::compareCluster()}}, \code{\link[DOSE:enrichDO]{DOSE::enrichDO()}}
}
