% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{enrich_gsea_kegg}
\alias{enrich_gsea_kegg}
\title{KEGG Gene Set Enrichment Analysis}
\usage{
enrich_gsea_kegg(
  dea_res,
  rank_by = "signed_log10p",
  aggr = "median",
  organism = "hsa",
  p_adj_method = "BH",
  p_cutoff = 0.05,
  min_gs_size = 10,
  max_gs_size = 500,
  exponent = 1,
  eps = 1e-10,
  seed = TRUE
)
}
\arguments{
\item{dea_res}{Differential analysis result. Can be one of:
\itemize{
\item Result from \code{\link[glystats:gly_limma]{glystats::gly_limma()}} (two groups), \code{\link[glystats:gly_ttest]{glystats::gly_ttest()}}, or \code{\link[glystats:gly_wilcox]{glystats::gly_wilcox()}},
called on an \code{\link[glyexp:experiment]{glyexp::experiment()}} of "traitproteomics" type.
\item A tibble with the following columns:
\itemize{
\item \code{protein}: Uniprot ID of proteins
\item \code{trait}: A glycosylation trait (e.g. "TFc" for proportion of core-fucosylated glycans)
\item \code{site}: The glycosylation site.
\item \code{p_val}: p-values, preferably adjusted p-values
\item \code{log2fc}: log2 of fold change
}
}}

\item{rank_by}{Criteria for ranking proteins.
One of the following:
\itemize{
\item "log2fc": log2 fold change with signs
\item "abs_log2fc": absolute log2 fold change
\item "log10p": negative log10 p-value
\item "signed_log10p" (default): log10 p-value with signs of log2 fold change
\item "log2fc_log10p": log2 fold change multiplied by negative log10 p-value
}}

\item{aggr}{Aggregation method for combining multiple scores across different traits and sites for the same protein.
One of "median", "mean", or "max". Defaults to "median".}

\item{organism}{KEGG organism code. Defaults to "hsa" (Homo sapiens).
See \code{\link[clusterProfiler:gseKEGG]{clusterProfiler::gseKEGG()}} for details.}

\item{p_adj_method}{P-value adjustment method.
One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".
Passed to \code{pAdjustMethod} of downstream enrichment function.
Defaults to "BH".}

\item{p_cutoff}{P-value cutoff to filter significant terms.
Passed to \code{pvalueCutoff} of downstream enrichment function. Defaults to 0.05.}

\item{min_gs_size}{Minimal size of each gene set for analyzing.
Gene sets with fewer genes than this threshold will be excluded.
Passed to \code{minGSSize} of downstream enrichment function. Defaults to 10.}

\item{max_gs_size}{Maximum size of each gene set for analyzing.
Gene sets with more genes than this threshold will be excluded.
Passed to \code{maxGSSize} of downstream enrichment function. Defaults to 500.}

\item{exponent}{Weight of each step. Passed to \code{exponent} of \code{\link[clusterProfiler:gseGO]{clusterProfiler::gseGO()}}.
Defaults to 1.}

\item{eps}{Epsilon for calculating p-values. Passed to \code{eps} of \code{\link[clusterProfiler:gseGO]{clusterProfiler::gseGO()}}.
Defaults to 1e-10.}

\item{seed}{Logical indicating whether to set a random seed for reproducibility.
Passed to \code{seed} of \code{\link[clusterProfiler:gseGO]{clusterProfiler::gseGO()}}. Defaults to \code{TRUE}.}
}
\value{
A clusterProfiler \code{gseaResult} object.
It can be readily converted to a tibble with \code{\link[tibble:as_tibble]{tibble::as_tibble()}},
or visualized with \code{clusterProfiler} functions like \code{\link[clusterProfiler:reexports]{clusterProfiler::ridgeplot()}}.
}
\description{
Performs KEGG pathway Gene Set Enrichment Analysis (GSEA)
on glycoproteins with dysregulated glycosylation.
}
\section{How it ranks proteins}{
GSEA requires a ranked list of proteins as input.
This function ranks proteins based on the median absolute log2 fold change across all traits and sites.
This reflects the overall glycosylation dysregulation degree of each glycoprotein.
You can use \code{rank_by} to specify other ranking criteria, such as p-values or signed log2 fold changes.
You can also use \code{aggr} to specify how to aggregate multiple scores for the same protein across different traits and sites.
}

\section{Common usage pattern}{
A common pattern of using this function is:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# 1. Perform differential analysis with `glystats`.
dea_res <- gly_ttest(exp)

# 2. Use this function.
go_res <- enrich_gsea_go(dea_res)  # or `enrich_gsea_xxx()` functions
}\if{html}{\out{</div>}}
}

\seealso{
\code{\link[clusterProfiler:gseKEGG]{clusterProfiler::gseKEGG()}}
}
