% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora.R
\name{enrich_ora_reactome}
\alias{enrich_ora_reactome}
\title{Reactome Over Representation Analysis}
\usage{
enrich_ora_reactome(
  dea_res,
  dea_p_cutoff = 0.05,
  dea_log2fc_cutoff = c(-1, 1),
  orgdb = "org.Hs.eg.db",
  organism = "human",
  universe = NULL,
  p_adj_method = "BH",
  p_cutoff = 0.05,
  q_cutoff = 0.2
)
}
\description{
Reactome Over Representation Analysis
}
