% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyfun-package.R
\docType{package}
\name{glyfun-package}
\alias{glyfun}
\alias{glyfun-package}
\title{glyfun: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://glycoverse.github.io/glyfun/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
