% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{detected_universe}
\alias{detected_universe}
\title{Helper function to prepare the \code{universe} parameter}
\usage{
detected_universe(x)
}
\arguments{
\item{x}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} or a \code{glystats} result.}
}
\value{
A character vector of protein UniProt IDs.
}
\description{
This function extracts all detected proteins in a \code{\link[glyexp:experiment]{glyexp::experiment()}} or a \code{glystats} result.
It can be readily passed to the \code{universe} parameter of all \code{glyfun} functions.
}
\examples{
library(glyexp)
universe <- detected_universe(real_experiment)
length(universe)
universe[1:5]
}
