% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora.R
\name{enrich_ora_reactome}
\alias{enrich_ora_reactome}
\title{Reactome Over Representation Analysis}
\usage{
enrich_ora_reactome(
  dea_res,
  dea_p_cutoff = 0.05,
  dea_log2fc_cutoff = c(-1, 1),
  organism = "human",
  universe = NULL,
  p_adj_method = "BH",
  p_cutoff = 0.05,
  q_cutoff = 0.2,
  min_gs_size = 10,
  max_gs_size = 500
)
}
\arguments{
\item{dea_res}{Differential analysis result. Can be one of:
\itemize{
\item Result from \code{\link[glystats:gly_limma]{glystats::gly_limma()}} (two groups), \code{\link[glystats:gly_ttest]{glystats::gly_ttest()}}, or \code{\link[glystats:gly_wilcox]{glystats::gly_wilcox()}},
called on an \code{\link[glyexp:experiment]{glyexp::experiment()}} of "traitproteomics" type.
\item A tibble with the following columns:
\itemize{
\item \code{protein}: Uniprot ID of proteins
\item \code{trait}: A glycosylation trait (e.g. "TFc" for proportion of core-fucosylated glycans)
\item \code{site}: The glycosylation site.
\item \code{p_val}: p-values, preferably adjusted p-values
\item \code{log2FC}: log2 of fold change
}
}}

\item{dea_p_cutoff}{P-value cutoff for statistical significance. Defaults to 0.05.
For \code{glystats} result input, adjusted p-values are used.}

\item{dea_log2fc_cutoff}{Log2 fold change cutoff statistical significance.
A length-2 numeric vector, being negative and positive boundaries, respectively.
For example, \code{c(-1, 1)} means "log2FC < -1 or log2FC > 1", and \code{c(-Inf, 1)} means "log2FC > 1".
Defaults to \code{c(-1, 1)}.}

\item{organism}{Reactome organism name. Passed to \code{organism} of \code{\link[ReactomePA:enrichPathway]{ReactomePA::enrichPathway()}}.
One of "human", "rat", "mouse", "celegans", "yeast", "zebrafish", "fly". Defaults to "human".}

\item{universe}{Background genes Uniprot IDs, directly passed to \code{universe} of downstream enrichment function.
If \code{NULL} (default), all genes in the data will be used.
Another common pattern is to use all detected proteins as backgroud genes.
You can use \code{\link[=detected_universe]{detected_universe()}} to help you.}

\item{p_adj_method}{P-value adjustment method.
One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".
Passed to \code{pAdjustMethod} of downstream enrichment function.
Defaults to "BH".}

\item{p_cutoff}{P-value cutoff to filter significant terms.
Passed to \code{pvalueCutoff} of downstream enrichment function. Defaults to 0.05.}

\item{q_cutoff}{Q-value (FDR) cutoff to filter significant terms.
Passed to \code{qvalueCutoff} of downstream enrichment function. Defaults to 0.2.}

\item{min_gs_size}{Minimal size of each gene set for analyzing.
Gene sets with fewer genes than this threshold will be excluded.
Passed to \code{minGSSize} of downstream enrichment function. Defaults to 10.}

\item{max_gs_size}{Maximum size of each gene set for analyzing.
Gene sets with more genes than this threshold will be excluded.
Passed to \code{maxGSSize} of downstream enrichment function. Defaults to 500.}
}
\value{
A clusterProfiler \code{enrichResult} object.
It can be readily converted to a tibble with \code{\link[tibble:as_tibble]{tibble::as_tibble()}},
or visualized with \code{clusterProfiler} functions like \code{\link[clusterProfiler:reexports]{clusterProfiler::dotplot()}}.
}
\description{
Performs Reactome pathway Over-Representation Analysis (ORA)
on glycoproteins with dysregulated glycosylation.
}
\section{Common usage pattern}{
A common pattern of using this function is:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# 1. Perform differential analysis with `glystats`.
dea_res <- gly_ttest(exp)

# 2. Use this function.
go_res <- enrich_gc_ora_go(dea_res)  # or other glyfun functions
}\if{html}{\out{</div>}}
}

\seealso{
\code{\link[ReactomePA:enrichPathway]{ReactomePA::enrichPathway()}}
}
